Struct quandl::quandl_request::QuandlRequest

pub struct QuandlRequest {
    // some fields omitted
}

Parameters for the request to Quandl API

Methods

impl QuandlRequest

fn new(database_code: &str, dataset_code: &str) -> QuandlRequest

Creates a new QuandlRequest using the specified database_code and dataset code. All other parameters as taken from the default implementation, setting the optional parameters to None.

fn api_key(self, key: String) -> QuandlRequest

Quandl API key. Used for premium databases and/or increased usage limits.

fn limit(self, limit: u64) -> QuandlRequest

You can use limit=n to get only the first n rows of your dataset. Use limit=1 to get the latest observation for any dataset.

fn rows(self, rows: u64) -> QuandlRequest

You can use rows=n to get only the first n rows of your dataset. Use rows=1 to get the latest observation for any dataset.

fn column_index(self, indx: u64) -> QuandlRequest

Request a specific column by passing the column_index=n parameter. Column 0 is the date column and is always returned. Data begins at column 1.

fn start_date<T: ?Sized + DateInput>(self, date: &T) -> Result<QuandlRequest>

Retrieve data within a specific date range, by setting start date for your query. Takes either a &str in the format of yyyy-mm-dd or a chrono::NaiveDate as input.

fn end_date<T: ?Sized + DateInput>(self, date: &T) -> Result<QuandlRequest>

Retrieve data within a specific date range, by setting end date for your query. Takes either a &str in the format of yyyy-mm-dd or a chrono::NaiveDate as input.

fn order(self, order: Order) -> QuandlRequest

Select the sort order. The default sort order is Desc.

fn collapse(self, collapse: Collapse) -> QuandlRequest

Parameters to indicate the desired frequency. When you change the frequency of a dataset, Quandl returns the last observation for the given period. By collapsing a daily dataset to monthly, you will get a sample of the original dataset where the observation for each month is the last data point available for that month. Set collapse with: collapse=none|daily|weekly|monthly|quarterly|annual.

fn transform(self, transform: Transform) -> QuandlRequest

Perform calculations on your data prior to downloading. The transformations currently available are row-on-row change, percentage change, cumulative sum, and normalize (set starting value at 100). Set the transform parameter with: transform=none|diff|rdiff|cumul|normalize.

fn run(&self) -> Result<Value>

Make a request to the Quandl API with the specified parameters