Struct quandl::quandl_request::QuandlRequest

pub struct QuandlRequest<'a> {
    pub quandl: &'a Quandl,
    pub database_code: String,
    pub dataset_code: String,
    pub limit: Option<u64>,
    pub rows: Option<u64>,
    pub column_index: Option<u64>,
    pub start_date: Option<NaiveDate>,
    pub end_date: Option<NaiveDate>,
    pub order: Option<Order>,
    pub collapse: Option<Collapse>,
    pub transform: Option<Transform>,
}

Parameters for the request to Quandl API

Fields
quandl: &'a Quandl

Reference to Quandl struct. This information will be available for all Quandl requests.

database_code: String

The unique database code on Quandl (ex. WIKI)

dataset_code: String

The unique dataset code on Quandl (ex. APPL)

limit: Option<u64>

You can use limit=n to get only the first n rows of your dataset. Use limit=1 to get the latest observation for any dataset.

rows: Option<u64>

You can use rows=n to get only the first n rows of your dataset. Use rows=1 to get the latest observation for any dataset.

column_index: Option<u64>

Request specific column.

start_date: Option<NaiveDate>

Retrieve data within a specific date range, by setting start dates for your query. Set the start date with: start_date=yyyy-mm-dd

end_date: Option<NaiveDate>

Retrieve data within a specific date range, by setting end dates for your query. Set the end date with: end_date=yyyy-mm-dd

order: Option<Order>

Sort in ascending or descending order.

collapse: Option<Collapse>

Parameters to indicate the desired frequency.

transform: Option<Transform>

Perform calculations on your data prior to downloading.

Methods

impl<'a> QuandlRequest<'a>

pub fn limit(self, limit: u64) -> QuandlRequest<'a>

Creates a new QuandlRequest using the specified database_code and dataset code. All other parameters as taken from the default implementation, setting the optional parameters to None. You can use limit=n to get only the first n rows of your dataset. Use limit=1 to get the latest observation for any dataset.

pub fn rows(self, rows: u64) -> QuandlRequest<'a>

You can use rows=n to get only the first n rows of your dataset. Use rows=1 to get the latest observation for any dataset.

pub fn column_index(self, indx: u64) -> QuandlRequest<'a>

Request a specific column by passing the column_index=n parameter. Column 0 is the date column and is always returned. Data begins at column 1.

pub fn start_date<T: ?Sized + DateInput>(
self,
date: &T
) -> Result<QuandlRequest<'a>>

Retrieve data within a specific date range, by setting start date for your query. Takes either a &str in the format of yyyy-mm-dd or a chrono::NaiveDate as input.

pub fn end_date<T: ?Sized + DateInput>(
self,
date: &T
) -> Result<QuandlRequest<'a>>

Retrieve data within a specific date range, by setting end date for your query. Takes either a &str in the format of yyyy-mm-dd or a chrono::NaiveDate as input.

pub fn order(self, order: Order) -> QuandlRequest<'a>

Select the sort order. The default sort order is Desc.

pub fn collapse(self, collapse: Collapse) -> QuandlRequest<'a>

Parameters to indicate the desired frequency. When you change the frequency of a dataset, Quandl returns the last observation for the given period. By collapsing a daily dataset to monthly, you will get a sample of the original dataset where the observation for each month is the last data point available for that month. Set collapse with: collapse=none|daily|weekly|monthly|quarterly|annual.

pub fn transform(self, transform: Transform) -> QuandlRequest<'a>

Perform calculations on your data prior to downloading. The transformations currently available are row-on-row change, percentage change, cumulative sum, and normalize (set starting value at 100). Set the transform parameter with: transform=none|diff|rdiff|cumul|normalize.

pub fn run(&self) -> Result<JsonValue>

Make a request to the Quandl API with the specified parameters

pub fn default(quandl: &'a Quandl) -> QuandlRequest<'a>

Create a default QuandlRequest